pyproject.toml
[tool.poetry]
name = "url-shortener-drf"
version = "0.1.0"
description = ""
authors = ["PeterChen <sanpeter992000@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "^5.1.7"
djangorestframework = "^3.15.2"
psycopg2 = "^2.9.10"
drf-yasg = "^1.21.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
